+++ golem/gui/static/js/test_common.js
@@ -151,4 +151,16 @@ var TestCommon = new function(){
              });
         };
     }
+
+    this.Utils = new function(){
+
+        this.convertTabToSpaces = function(cm) {
+            if (cm.somethingSelected()) {
+                cm.indentSelection("add");
+            } else {
+                cm.replaceSelection(cm.getOption("indentWithTabs")? "\t":
+                Array(cm.getOption("indentUnit") + 1).join(" "), "end", "+input");
+            }
+        }
+    }
 }